+++ src/main/kotlin/calculator/Splitted.kt
@@ -1,0 +1,6 @@
+package calculator
+
+data class Splitted(private val elements: List<String>) : List<String> by elements {
+    constructor(expression: Expression) : this(expression.syntax(), expression.delimiter)
+    constructor(string: String, delimiter: String) : this(string.split(delimiter))
+}
