+++ src/main/kotlin/calculator/Sum.kt
@@ -1,0 +1,5 @@
+package calculator
+
+data class Sum(private val intList: List<Int>) {
+    fun toInt() = intList.sum()
+}
