+++ src/test/kotlin/calculator/IntegrationTest.kt
@@ -1,9 +1,5 @@
 package calculator
 
-import calculator.ExpressionTest.Expression
-import calculator.NonNegativeIntListTest.NonNegativeIntList
-import calculator.SplittedTest.Splitted
-import calculator.SumTest.Sum
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
