+++ src/test/kotlin/calculator/NonNegativeIntListTest.kt
@@ -20,18 +20,4 @@ class NonNegativeIntListTest {
     fun `숫자가 아니면 예외가 발생한다`() {
         assertThrows<RuntimeException> { NonNegativeIntList.of(listOf("1", "a")) }
     }
-
-    data class NonNegativeIntList(private val nonNegativeIntList: List<Int>) : List<Int> by nonNegativeIntList {
-        init {
-            require(nonNegativeIntList.find { it < 0 } == null)
-        }
-
-        companion object {
-            fun of(numericStringList: List<String>): NonNegativeIntList {
-                return numericStringList.map {
-                    it.toInt()
-                }.let { NonNegativeIntList(it) }
-            }
-        }
-    }
 }
