+++ src/test/kotlin/calculator/SumTest.kt
@@ -8,8 +8,4 @@ class SumTest {
     fun `숫자 리스트를 모두 더한다`() {
         assertThat(Sum(listOf(1, 2)).toInt()).isEqualTo(3)
     }
-
-    data class Sum(private val intList: List<Int>) {
-        fun toInt() = intList.sum()
-    }
 }
